# packages/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: TODOMaster API
  description: |
    TODOMaster是一个基于OKR（Objectives and Key Results）方法论的任务管理系统。
    本API提供了完整的用户管理、目标设置、关键结果跟踪和任务管理功能。
  version: 1.0.0
  contact:
    name: TODOMaster Team
    email: support@todomaster.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: 开发服务器
  - url: https://api.todomaster.com/v1
    description: 生产服务器
...
security:
  - bearerAuth: []
paths:
...
  /objectives:
    get:
      tags: [objectives]
      summary: 获取目标列表
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: status
          in: query
          schema:
            type: string
            enum: [active, completed, archived]
        - name: category
          in: query
          schema:
            type: string
            enum: [work, personal, learning, health]
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 目标列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectivesListResponse'

    post:
      tags: [objectives]
      summary: 创建目标
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateObjectiveRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectiveDetailResponse'
  /objectives/{id}:
    get:
      tags: [objectives]
      summary: 获取目标详情
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 目标详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectiveDetailResponse'

    put:
      tags: [objectives]
      summary: 更新目标
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateObjectiveRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectiveDetailResponse'

    delete:
      tags: [objectives]
      summary: 删除目标
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
...
  /stats/objectives:
    get:
      tags: [stats]
      summary: 获取目标统计数据
      security:
        - bearerAuth: []
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 目标统计数据
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectiveStatsResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Objective:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [work, personal, learning, health]
        status:
          type: string
          enum: [active, completed, archived]
        progress:
          type: number
          minimum: 0
          maximum: 100
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        userId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer

    # 请求模型
...
    CreateObjectiveRequest:
      type: object
      required:
        - title
        - category
        - startDate
        - endDate
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [work, personal, learning, health]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
    UpdateObjectiveRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, completed, archived]
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        message:
          type: string
...
    SuccessMessage:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
...
    ObjectivesListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                objectives:
                  type: array
                  items:
                    $ref: '#/components/schemas/Objective'
                pagination:
                  $ref: '#/components/schemas/Pagination'
    ObjectiveDetailResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                objective:
                  $ref: '#/components/schemas/Objective'
...
    ObjectiveStatsResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                summary:
                  type: object
                statusDistribution:
                  type: object
                categoryDistribution:
                  type: object
                progressDistribution:
                  type: object
...